feat: ros2 integration
- Add ros_bridge ROS2 package (robot_node, mqtt_bridge, launch files)
- Add POST /robots/{id}/command endpoint for MQTT command publishing
- Update schemas and models to use float for x/y/battery
- Add Alembic migration for float columns
- Fix frontend to use command endpoint for STOP/RESUME/CHARGE
- Fix WebSocket to handle new robots appearing live
- Add ENABLE_PYTHON_SIM toggle to backend
- Update .gitignore for ROS2 build artifacts

+++ backend/app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime
+from sqlalchemy import Column, Float, Integer, String, DateTime
 from sqlalchemy.sql import func
 from .database import Base
 
@@ -8,9 +8,9 @@ class Robot(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String, index=True)
-    battery = Column(Integer)
+    battery = Column(Float)
     status = Column(String)
-    x = Column(Integer)
-    y = Column(Integer)
+    x = Column(Float)
+    y = Column(Float)
     task = Column(String, nullable=True)
     last_heartbeat = Column(DateTime(timezone=True), server_default=func.now())
